Extract mouse UI overlay support services
Move frame waiting, overlay animation, and main-thread cleanup scheduling into cohesive helpers while preserving timing, cancellation, and cleanup behavior. Remove the one-line overlay wrapper and local animation constant aliases so callers use the owning APIs directly.
